tests restructure
